Disable selecting incompatible storage classes (and move disabled item CSS from ProviderSelect to SimpleSelect)
