--- src/app/common/components/ProviderSelect/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ProviderSelect';
