+++ src/app/common/components/SimpleSelect/index.ts
@@ -1,0 +1,2 @@
+export * from './SimpleSelect';
+export { default } from './SimpleSelect';
